+++ packages/substrate/src/wanxiang_substrate/worldpack/__init__.py
@@ -1,0 +1,21 @@
+"""RedChamber WorldPack assembly substrate (G35I)."""
+
+from wanxiang_substrate.worldpack.assembler import (
+    AssembledWorldPack,
+    GenesisSpec,
+    WorldPackAssembler,
+    WorldPackSignature,
+    literary_constitution,
+    sign_world_pack,
+    verify_world_pack,
+)
+
+__all__ = [
+    "AssembledWorldPack",
+    "GenesisSpec",
+    "WorldPackAssembler",
+    "WorldPackSignature",
+    "literary_constitution",
+    "sign_world_pack",
+    "verify_world_pack",
+]